CORE-12120 : Review and implement missing contract rules. (#18)
**Reviewed all contract constraints, added where missing, removed where unnecessary, and updated all associated documentation.**

Some of the contract commands were checking for creation or consumption of states that weren't part of the associated command's remit; for example: `ChainableConstraints.verifyCreate` (which is called by `ChainableContractCreateCommand`) was checking that:

> On chainable state(s) creating, zero chainable states must be consumed.

It doesn't make sense to perform this check here, since the command deals with creation of `ChainableState` instances, and therefore we should not care whether any are consumed. Checking for state consumption on creation, and state creation on deletion should be concerns that are implemented by CorDapp developers, and not enforced by us.

As such, all contracts now abide by the following rules:

 - **Create** commands check that at least one of the associated state has been **created**.
 - **Update** commands check that at least one of the associated state has been **consumed**, and **created**.
 - **Delete** commands check that at least one of the associated state has been **consumed**.

**This PR also includes the changes required for CORE-11992 (Review base rules for identifiable commands).** 

As such, the behaviour of the identifiable constraints has changed slightly. Previously, `IdentifiableContract` checked that any `IdentifiableState` output with a non-null identifier must consume a state where that identifier is either the consumed state's `StateRef`, or its identifier. Realistically, this only makes sense when updating an `IdentifiableState` and as such the constraint has been refactored to apply only to the `IdentifiableConstraints.verifyUpdate` (called by `IdentifiableContractUpdateCommand`).

Additional rules have been specified for:
 - `IdentifiableConstraints.verifyCreate` to ensure that at least one `IdentifiableState` has been created.
 - `IdentifiableConstraints.verifyDelete` to ensure that at least one `IdentifiableState` has been deleted.

(thus aligning with the comments, above)