CORE-17403 Handle stable vault-named query changes
Updates in 5.1 broke the usage of `setOffset` as the stable query paging
of vault named queries was added.

Changes to some of the ledger extension method and query tests have been
updated to handle this change.

Also includes changes to bring the e2e tests more in line with those in
the e2e test repo (mainly notary version changes).