CORE-14241 Fix the logic in `CONTRACT_RULE_UPDATE_IDENTIFIER_EXCLUSIVITY MISSING_ID` (#56)
Fix the logic in the test flow's CONTRACT_RULE_UPDATE_IDENTIFIER_EXCLUSIVITY MISSING_ID branch. We now generate a new "fake ID" and add it to the outputs list. That way it will 100% not be present in the inputs list. This way triggering CONTRACT_RULE_UPDATE_IDENTIFIERS constraint.
Enable the test in IdentifiableTests